started creating the add order functionality
currently the orders to add are being displayed by the enpoint, now adding of the orders needs to be implemented

+++ src/main/java/solutional/homework/ecommerce/Models/DTO/OrderResponseDTO.java
@@ -40,7 +40,7 @@ public class OrderResponseDTO {
         private int quantity;
         private UUID replacedWith;
 
-        public ProductDTO(Integer id, String name, String price, int quantity) {
+        public ProductDTO(Long id, String name, String price, int quantity) {
         }
     }
 }
